feat: add favorite tag label details page
